Implements the DelegateProxy
Includes tests for void delegate methods. Now, rather than having to
check whether the delegate responds to a given selector you can just
send it. An exception will be thrown if you attempt to call a method
which isn't on any of the protocols the delegate conforms to,
irrespective of whether or not an implementation exists.